refactor(config): introduce a new layered config system (#680)
This commit introduces a new configuration system which aims to be
centralized and easy to comprehend while retaining a certain amount of
flexibility (loading parts of it interactively, e.g. when switching
channels, etc.).

NOTE: I did the field merging by hand but I'm sure we can simplify it
further with crates such as `merge` now that the overall logic is in
place.

## 📺 PR Description

### Overview
The basic idea is simple, we now have a unique struct from which we read
configuration values throughout the application:
```rs
pub struct MergedConfig {...}
```
This holds all of tv's configuration options, regardless of where they
come from (the cli, the user's config, channels).

This struct can only be constructed from a `LayeredConfig`, which holds
tv's different configuration layers:
```rs
pub struct LayeredConfig {
    base_config: Config,
    channel: ChannelPrototype,
    /// things passed through the CLI that only apply to the first channel loaded
    channel_cli: ChannelCli,
    /// things passed through the CLI that apply for the entire session
    global_cli: GlobalCli,
}
```
```rs
let merged_config = layered_config.merge();
```

The program keeps both the layered and the merged config in memory so
that whenever we wish to update a layer of the config (e.g. when
changing channels to a new prototype and all its associated
configuration) we can simply do so by updating the layered config and
producing a new merged one.

Here's a diagram of the general idea:
```mermaid
  graph TB
      subgraph "Configuration Sources"
          A[Embedded Defaults]
          B[User Config<br/>~/.config/television/config.toml]
          C[Channel Config<br/>cable/*.toml]
          D[CLI Arguments<br/>--flags]
      end

      subgraph LayeredConfig
          E[Base Config]
          F[Channel Prototype]
          G[CLI Args]
      end

      F --|overrides|--> E
      G --|overrides|--> F

      A --> E
      B --> E
      C --> F
      D --> G

      LayeredConfig --> H[MergedConfig]
```

## Checklist

<!-- a quick pass through the following items to make sure you haven't
forgotten anything -->

- [x] my commits **and PR title** follow the [conventional
commits](https://www.conventionalcommits.org/en/v1.0.0/#summary) format
- [ ] if this is a new feature, I have added tests to consolidate the
feature and prevent regressions
- [ ] if this is a bug fix, I have added a test that reproduces the bug
(if applicable)
- [x] I have added a reasonable amount of documentation to the code
where appropriate

+++ television/event.rs
@@ -90,7 +90,7 @@ pub enum Key {
 /// use television::event::{InputEvent, Key, MouseInputEvent};
 /// use television::keymap::InputMap;
 ///
-/// let input_map = InputMap::new();
+/// let input_map = InputMap::default();
 ///
 /// // Handle keyboard input
 /// let key_event = InputEvent::Key(Key::Enter);
@@ -257,9 +257,9 @@ fn flush_existing_events() {
 }
 
 impl EventLoop {
-    pub fn new(tick_rate: f64) -> Self {
+    pub fn new(tick_rate: u64) -> Self {
         let (tx, rx) = mpsc::unbounded_channel();
-        let tick_interval = Duration::from_secs_f64(1.0 / tick_rate);
+        let tick_interval = Duration::from_secs_f64(1.0 / tick_rate as f64);
 
         let (abort, mut abort_recv) = mpsc::unbounded_channel();
